+++ packages/react/src/components/Table/TableToolbar/TableToolbarSVGButton.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 
@@ -16,20 +16,61 @@ const propTypes = {
   isActive: PropTypes.bool,
   /** is the button disabled */
   disabled: PropTypes.bool,
+  tooltipAlignment: PropTypes.oneOf(['start', 'center', 'end']),
+  tooltipPosition: PropTypes.oneOf(['top', 'right', 'bottom', 'left']),
+  langDir: PropTypes.oneOf(['ltr', 'rtl']),
 };
 
 const defaultProps = {
   isActive: false,
   disabled: false,
+  tooltipAlignment: 'center',
+  tooltipPosition: 'top',
+  langDir: 'ltr',
 };
 
 /**
  * Toolbar button that renders an icon only button
  */
-const TableToolbarSVGButton = ({ onClick, testId, className, description, isActive, ...rest }) => {
+const TableToolbarSVGButton = ({
+  onClick,
+  testId,
+  className,
+  description,
+  isActive,
+  tooltipAlignment,
+  tooltipPosition,
+  langDir,
+  ...rest
+}) => {
+  const [adjustedTooltipAlignment, setAdjustedTooltipAlignment] = useState(tooltipAlignment);
+
   return (
     <Button
       {...rest}
+      ref={(node) => {
+        /* istanbul ignore else */
+        if (node) {
+          const tooltip = node.querySelector(`.${prefix}--assistive-text`);
+          if (!tooltip) {
+            return;
+          }
+          const childRect = tooltip.getBoundingClientRect();
+          const parentRect = node.parentElement.getBoundingClientRect();
+          /* istanbul ignore else */
+          if (
+            langDir === 'ltr' &&
+            childRect.left + childRect.width > parentRect.left + parentRect.width
+          ) {
+            setAdjustedTooltipAlignment('end');
+            return;
+          }
+          /* istanbul ignore else */
+          if (langDir === 'rtl' && childRect.left < parentRect.left) {
+            setAdjustedTooltipAlignment('start');
+          }
+        }
+      }}
       className={classnames(
         `${prefix}--btn--icon-only`,
         `${iotPrefix}--tooltip-svg-wrapper`,
@@ -38,11 +79,12 @@ const TableToolbarSVGButton = ({ onClick, testId, className, description, isActi
           [`${iotPrefix}--table-toolbar-button-active`]: isActive, // https://github.com/carbon-design-system/carbon/issues/6160
         }
       )}
-      kind="ghost"
+      kind="icon-selection"
       onClick={onClick}
-      title={description}
       iconDescription={description}
       testId={testId}
+      tooltipAlignment={adjustedTooltipAlignment}
+      tooltipPosition={tooltipPosition}
     />
   );
 };